+++ CHANGELOG.md
@@ -71,6 +71,7 @@ Advantages:
 - [thermostat]: Conformance to new matter.js internal attributes externalMeasuredIndoorTemperature and externallyMeasuredOccupancy (not in Matter 1.4.2 specs).
 - [scenes]: Added SceneManagement when required.
 - [server]: Added start and stop devices server mode in plugin restart.
+- [server]: Added start and stop devices server mode in plugin enable and disable.
 - [matter.js]: Bump to matter.j v. 0.16.0.
 - [matter.js]: Bump to matter.j v. 0.16.1.
 - [matter.js]: Bump to matter.j v. 0.16.2.